// src/app/core/app/enum/SiteSummary.enum.ts
enum SiteSummary {
  MAX_LEVEL = 'Niveau max : ',
  MIN_LEVEL = 'Niveau mini : ',
  ROCK_TYPE = 'Type de rocher : ',
  EXPOSITION = 'Exposition : ',
  ROUTE_NUMBER = 'Nombre de voies : ',
  ROUTE_HEIGHT = 'Hauteur des voies : ',
  ROUTE_PROFILE = 'Profil des voies : ',
  EQUIPMENT = 'Equipements : ',
}
export { SiteSummary };